Backend ui jc before presentation

+++ DoConnect/DocConnectApp/App/Controllers/PracticesController.js
@@ -1,25 +1,21 @@
-﻿app.controller("PracticesController", ["$scope", "PracticesService", "$interval", 
-    function ($scope, PracticesService, $interval) {
+﻿app.controller("PracticesController", ["$scope", "PracticesService", "$interval", "$ngBootbox",
+    function ($scope, PracticesService, $interval, $ngBootbox) {
 
         $scope.PageTitle_Practices = 'Practices';
         $scope.PageTitle_NewPractice = 'New Practices Details';
-        
-        //Sort Function
+
         $scope.sort = function (keyname) {
             $scope.sortKey = keyname;
             $scope.reverse = !$scope.reverse;
         };
 
-        //Select All Practices
         $scope.GetAllPractices = function () {
-            PracticesService.GetAllPractices().then
-            (function (result) {
+            PracticesService.GetAllPractices().then(function (result) {
                 $scope.Practices = result.data;
             });
         };
         $scope.GetAllPractices();
 
-        //Select PracticeByID Function
         $scope.ViewPractice = function (ID) {
             PracticesService.GetPracticeByID(ID).success(function (result) {
                 $scope.ID = result["ID"];
@@ -34,7 +30,6 @@
             });
         };
 
-        //Insert Practice Funtion
         $scope.NewPractice = function (Name, Phone_Number, Fax_Number, Street_Address, Suburb, City, Country, Trading_Times) {
             PracticesService.InsertPractice(Name, Phone_Number, Fax_Number, Street_Address, Suburb, City, Country, Trading_Times).success(function () {
                 $scope.GetAllPractices();
@@ -43,10 +38,9 @@
             },
                 function (error) {
                     btnAlert("System Error Message", "Insert unsuccessful.");
-            });
+                });
         };
 
-        //Update Practice Funtion
         $scope.function_btnUpdatePractice = function (ID) {
             var btnText = angular.element("#function_btnUpdatePractice").html();
             if (btnText == "Update") {
@@ -67,15 +61,25 @@
             }
         };
 
-        //Delete Practice Funtion
-        $scope.DeletePractice = function () {
-            //var DeleteStatus = btnConfirm("info", "Are you sure you want to delete this Practice?");
-            //if (DeleteStatus == "true") {
-                PracticesService.DeletePractice($scope.ID).then(function () {
-                    $scope.GetAllPractices();                                       
+        $scope.DeletePractice1 = function (ID, Name) {
+            $ngBootbox.confirm("Are you sure you want to delete this Practice: " + Name + " ?").then(function () {
+                PracticesService.DeletePractice(ID).then(function () {
+                    $scope.GetAllPractices();
+                    btnSuccess("Practice record successfully deleted.");
                 }, function (error) {
                     btnAlert("System Error Message", "Delete unsuccessful.");
                 });
-            //}            
+            }, function () { });
+        };
+        $scope.DeletePractice2 = function () {
+            $ngBootbox.confirm("Are you sure you want to delete this Practice: " + $scope.Name + " ?").then(function () {
+                PracticesService.DeletePractice($scope.ID).then(function () {
+                    $scope.GetAllPractices();
+                    angular.element("#CloseModel").trigger("click");
+                    btnSuccess("Practice record successfully deleted.");
+                }, function (error) {
+                    btnAlert("System Error Message", "Delete unsuccessful.");
+                });
+            }, function () { });
         };
     }]);
